added typescript and alert-logs-utils file
